feat(decomp): byte-match emi/world00/area030/04@0x801D2034

+++ src/emi/world00/area030/04/func_801D2034.c
@@ -6,5 +6,5 @@
 void func_801D2034(s16 arg0, s16 arg1, u8 arg2, s8 arg3) {
   func_801E0C80(0, arg2);
   func_801E0DCC(0x17, arg2, arg0, arg1);
-  func_801E0DCC((s16)(arg3 + 0x17), arg2, (s16)(arg0 + 0x28), arg1);
+  func_801E0DCC((u8)(arg3 + 0x17), arg2, (s16)(arg0 + 0x28), arg1);
 }
